feat: Added a server configuration which controls the assumption whether a Langauge Client interprets the text received in "insertText" of completion requests.
Some Language Clients, such as Kate, insert the received Text verbatim into the editor. Other clients like VS Code or Neovim match the received text against the word under the cursor and only insert the difference.

The description of the insertText field in the LSP spec provides an example:
https://microsoft.github.io/language-server-protocol/specifications/specification-3-17/#completionItem

By default, RST Language Server assumes that the client does interpret the insertText, which seems to be the more prevalent behaviour.